LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/popup-coop-by-sw-from-coop.https-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: BroadcastChannel
CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.

Harness Error (FAIL), message = ReferenceError: Can't find variable: BroadcastChannel
